Refactor Ref
