--- src/ref.js
@@ -1,5 +1,0 @@
-import {symbol} from "./shared.js";
-
-export const setRef = symbol();
-
-export const ref = symbol();
